requirements.txt
appdirs==1.4.3
attrs==19.3.0
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
mongoengine==0.23.1
more-itertools==8.4.0
msgpack==0.6.2
packaging==20.4
pep517==0.8.2
pip-autoremove==0.9.1
pluggy==0.13.1
progress==1.5
py==1.10.0
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
pytoml==0.1.21
requests==2.25.1
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.26.5
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
